Renamed seventv to emote
