+++ src/commands/fun/emote.ts
